#4 Automation project fails to re-open after it is closed.
When re-opening a automation project, compiled assembly name must be unique . Earlier assembly name sequence was getting reset causing same assembly name to be used which was already loaded. Hence, new assembly could not be saved on disk. This was already handled for Prefab Projects. Refactored method into based Project Manager class.
